first commit for gitops demo
The initial bases are setup and a little terraform. Still need to
generalize things a bit for each environment, and add some cluster
specific things to the bases/workloads such as external secrets,
replacements for things like region and possible environment variables.

Still a lot of work to do but wanted to check this in for now. Current
goals are to bootstrap the cluster with argocd, cert-manager,
external-secrets, and GitHub actions runner controller.

At that point I can build out some CICD stuff and run my applications.